backend/requirements.txt
# FastAPI和相关依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# OpenAI SDK (for VolcEngine API compatibility)
openai>=1.0.0

# 日志和配置
python-multipart==0.0.6
python-dotenv==1.0.0

# 音频处理 - 只保留numpy，移除pydub相关
numpy==1.24.3

# 工具类
typing-extensions==4.8.0
pydantic==2.5.0 